services/vector-store/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pinecone-client==3.0.0
openai==1.10.0
tenacity==8.2.3
pydantic==2.6.0
python-dotenv==1.0.0
